Add solution to Climbing the Tree
Ohh my! I have found the solution myself, then made 2 mistakes
which I haven't noticed for hours!
First that I didn't account for the posibility that the first message
could be a request instead of a report, leading to overflows.
Secondly, this wasn't that bad, but I didn't rearrange the equeations
describing the constraints and had to do a binary search instead.