Add AOC/2023/3 part one and two
Initially had some issue understanding the difference between runes
and bytes. In this case it is not the problem that I am mixing
runes and string indexing, because we are handling ASCII characters.
If this wasn't the case though, then SymbolAround should have to work
on a matrix of runes.
Ahh text encodings, so nice.