Add AOC/2023/21 part one solution
Part 2 looks tricky, but I see some patterns. The board gets filled
it will alternate between two values. This is the same with multiple
boards. What isn't entirely clear how does this extrapolate.
Upon investigating the 9x9, it can be said that the sum of the 9x9
will alternate between 4*the lower value + 5*the higher value, and the
4, 5 switched. This makes reasonable sense. Half the boards will be in
this state, while the other half in the other. Good. But how could
you tell how many partially filled boards will be there and far will
those be filled?
Have a hunch that either this can be seen from how one board gets filled.
Also looked at the patterns, with just the Os being visible and it reminded
me of a crystal growing, but this is probabbly irrelevant.