Add AOC/2023/22 part two solution
Quite happy about the elegance of my solution. What I don't like
however is that I had to copy my PriorityQueue code again and
worse I don't quite understand those interface interactions and
requirements it implements. It would be nice to have it in the standard.
As a Min, Max, Abs function too.