Add AOC/2023/18 part two solution
Not gonna lie, this nearly broke me. Not because it is particularly
difficult, no I almost immediatelly thought of the scanline algorithm
when seeing it. It is difficult because there are so many edge cases
and so many ways to approach the scanline solution with their own
edge cases. As you may see from commits I have tried at least 3 solutions.
For some reason I insisted on creating some bounding class and update
it while I scanned from top to button. It would update its width,
when meeting horizontal lines. This always worked correctly, but there
was no way to reasonable tell what was the width of such a transition line.
This would have a whole host of edge cases.
Then after 5 days of being broken and wasting a total of about 20
hours of trying to implement those solutions I have realized I have
been overcomplicating things all along.
From the beginning I have understood that we are within the shape
when between two vertical lines, but for hubris I never spent the five
minutes to write down all the cases to decide when to count and until
when. So I have never noticed there were only 3 options, and that if
the lines were ordered by their vertical position all I needed to
do was to go from left to right and if the vertical line wasn't marked
as an ending line I needed to find the next closes vertical, which then
I could mark as an ending one.
So elegant. Ahh, and I was stubborn trying to fix overcomplicated solutions.
Live and learn.