Add AOC/2024/17 part two
Not exstatic about the solution, but for now it is okay.
We properly crack the digits in order as we should, but it isn't
properly handled if a given cycle may not contain a certain digit
with the given configuration. This can easily happan as for each digit
place there will be a set of digits that may apper which depend on
previously set digits.
It can also happen that for a given digit position the same number
may appear multiple times in a cycle.