Add AoC/2024/24 part two solution
Well, not quite. Did the solution by hand, instead of coding up an
automatic one. Time is too precious you know.
The reason is that we know this has to be an optimal n digit adder and
as the comments describe we know how it needs to look like.
We know exactly what needs to happen and where. So from the first bit
we can easily track and correct all mistakes on the first go.
This however only works if you can identify exactly what do you see
when you see an issue. This is realy not trivial to describe in code,
but by doing it by hand it is relatively simple with the supplied
support functions.
Didn't like the solutions I have seen where it will simulate adding
numbers, find a faulty place then switch things around. They run slowly
and feel icky, so by hand it is.