Add AOC/2025/6 part 1&2 in Zig
Uhh this was a painful day for me.
Don't really understand the memory model of Zig, what is exactly copied,
what not, what can be modified what not.
How to add an item to a collection.
Jeezz I hope I am just using it wrong, because always having to access
ArrayList.items field to do anything with the collections seeems very
off.
Not to mention all the houplas one has to jump through to elements to
it.
Sheesh, and no string support, no shadowing, no easy access to iterators
or flexible for loops.
For example why can't I just write "for (lines)" where lines
is an ArrayList?
Why do I have to reach into the container for the data with
"for (lines.items)"?

The language is getting rather annoying with these weird conventions and
the severe lack of documentations.