Add AOC/2025/7 part 1&2 in Zig
It is so intuitive to use the ArrayList collection, that I am
not entirely sure it has a point beyond defining a collection of
functions.
Copying it only copies the most outer level. If the elements
are ArrayLists as well, those will become shared between the
two collections.
I don't think this is in accordance with Zig Zen.
If one calls clone on a collection explicitly, has to pass an
allocator ect, it should be expected that the whole thing is
copied deeply.