Add AOC/2025/9 part 2 in Zig
Dear almighty this almost broke me!
Ideally in these problems you convert the input to a matrix,
do a scanline, identify internal points, generate all rectangles,
check if they are inside with another scanline, pick biggest done.
Unfortuanatelly the space is 100k * 100k big so a matrix would take up
at least 10k MiB or 10 GiB, which is a bit too much for my liking.

So instead we just generate the internal vertical/horizontal spans when there is any
change in the shape. Both of these take about N (number of points)/2 * 3
* sizeof(i64) space which in our case, N = 500, should be around
250 * 3 * 8 = 6000 bytes.
With this we brute generate all possible rectangles, then check if
they are within the horizontal/vertical internal sections.
If so we pick the biggest.
This is technically an N^3 algorithm, which is a bit shit.
N^2 is for generating all possible rectangles, while another N is
for validating if its internal. In practice this final N should
be much faster because it finds the start in logn using binary search
then checks until the end.
In any case 500 nodes, 500^3 = 125*10^6 => (with 10^9 operations on a
single) core will take about 125 milliseconds.
Decently okay, granted the final memory usage and the final simplicity
of the algorithm.
Seen much faster solutions out there, but this is less complext.

PS.: I know some functions are almost duplicates, but I really didn't
want to complicate the code with additional selectors to be able to
A
reuse them. Good enough for now.